pyproject.toml
[tool.poetry]
name = "llm4kg"
version = "0.1.0"
description = "Framework for orchestrating Knowledge Graph Construction"
authors = ["Marvin Hofer"]
license = "MPL-2.0"
readme = "README.md"
packages = [{include = "llm4kg_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "1.12.0"
transformers = "^4.34.0"
torch = "^2.1.0"
accelerate = "^0.23.0"
pyspark = "^3.5.0" 
pyyaml = "^6.0.1"
pytest = "^7.4.4"
jsonlines = "^4.0.0"
rdflib = "7.0.0"
sparqlwrapper = "^2.0.0"
wptools = "^0.4.17"
schema = "^0.7.5"
pymongo = "^4.6.1"
anthropic = "0.17.0"
streamlit = "^1.31.1"
scikit-learn = "^1.4.1.post1"
#llm-kg-bench = {git = "git@github.com:AKSW/LLM-KG-Bench.git", rev = "dev"}
matplotlib = "^3.8.3"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
config_test = "llm4kg_core.config:run"
# llm4kg_mapping = "llm4kg_mapping.run:run"

